# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PMC Medical RAG API",
    "description": "基于 PMC oa_comm 医学文献的检索增强生成服务",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/qa": {
      "post": {
        "tags": [
          "qa"
        ],
        "summary": "Ask",
        "operationId": "ask_api_v1_qa_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QARequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_QAResponseData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/qa/stream": {
      "post": {
        "tags": [
          "qa"
        ],
        "summary": "Ask Stream",
        "operationId": "ask_stream_api_v1_qa_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QARequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/documents": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "List Documents",
        "operationId": "list_documents_api_v1_documents_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "页码，从 1 开始",
              "default": 1,
              "title": "Page"
            },
            "description": "页码，从 1 开始"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量，1-100",
              "default": 20,
              "title": "Page Size"
            },
            "description": "每页数量，1-100"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_PaginatedData_DocumentOut__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{doc_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get Document",
        "operationId": "get_document_api_v1_documents__doc_id__get",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_DocumentOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DocumentOut": {
        "properties": {
          "doc_id": {
            "type": "string",
            "title": "Doc Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "abstract": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abstract"
          },
          "journal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Journal"
          },
          "pub_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pub Date"
          },
          "pmid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pmid"
          },
          "doi": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doi"
          },
          "article_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Article Type"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          }
        },
        "type": "object",
        "required": [
          "doc_id",
          "title",
          "chunk_count"
        ],
        "title": "DocumentOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PageInfo": {
        "properties": {
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          }
        },
        "type": "object",
        "required": [
          "page",
          "page_size",
          "total",
          "total_pages"
        ],
        "title": "PageInfo"
      },
      "PaginatedData_DocumentOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentOut"
            },
            "type": "array",
            "title": "Items"
          },
          "page_info": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page_info"
        ],
        "title": "PaginatedData[DocumentOut]"
      },
      "QARequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query",
            "description": "用户问题，非空，最长 2000 字符"
          },
          "top_k": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K",
            "description": "最终返回结果数量，1-20",
            "default": 8
          },
          "fusion_strategy": {
            "type": "string",
            "enum": [
              "rrf",
              "weighted",
              "simple"
            ],
            "title": "Fusion Strategy",
            "description": "多路检索融合策略",
            "default": "rrf"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "会话 ID；传入则关联历史对话，不传则新建/不关联"
          },
          "run_evaluation": {
            "type": "boolean",
            "title": "Run Evaluation",
            "description": "是否执行证据评估阶段",
            "default": true
          },
          "run_review": {
            "type": "boolean",
            "title": "Run Review",
            "description": "是否执行批判性审查阶段",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "QARequest"
      },
      "QAResponseData": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceItem"
            },
            "type": "array",
            "title": "Sources"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "total_time_seconds": {
            "type": "number",
            "title": "Total Time Seconds"
          },
          "citation_retry_attempts": {
            "type": "integer",
            "title": "Citation Retry Attempts"
          },
          "format_check_pass": {
            "type": "boolean",
            "title": "Format Check Pass"
          }
        },
        "type": "object",
        "required": [
          "answer",
          "sources",
          "total_time_seconds",
          "citation_retry_attempts",
          "format_check_pass"
        ],
        "title": "QAResponseData"
      },
...
      "ResponseModel_DocumentOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocumentOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "title": "ResponseModel[DocumentOut]"
      },
...
      "ResponseModel_PaginatedData_DocumentOut__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginatedData_DocumentOut_"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "title": "ResponseModel[PaginatedData[DocumentOut]]"
      },
      "ResponseModel_QAResponseData_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QAResponseData"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "title": "ResponseModel[QAResponseData]"
      },
...
      "SourceItem": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id"
          },
          "journal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Journal"
          },
          "pub_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pub Year"
          },
          "pmc_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pmc Id"
          },
          "relevance_score": {
            "type": "number",
            "title": "Relevance Score"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "chunk_id",
          "relevance_score"
        ],
        "title": "SourceItem"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}